+++ frameimpl_test.go
@@ -26,7 +26,7 @@ import (
 )
 
 func TestFrameNew(t *testing.T) {
-	val0, val1, size := 7, "n", 10
+	val0, val1, size := int64(7), "n", 10
 	col0, _ := NewLabelColumn("col0", val0, size)
 	col1, _ := NewLabelColumn("col1", val1, size)
 	cols := []Column{col0, col1}
